requirements.txt
flask
gunicorn
google-generativeai
psycopg2-binary 